scripts/requirements.txt
openai>=1.35.0
requests>=2.31.0
pyyaml>=6.0
